+++ phase2/src/ATM/Accounts/Depositable.java
@@ -3,5 +3,5 @@ package ATM.Accounts;
  * The interface for depositable of the accounts
  */
 public interface Depositable {
-    void deposit(double amount);
+    void deposit(int amount);
 }
