+++ phase2/src/ATM/Accounts/Withdrawable.java
@@ -2,5 +2,5 @@ package ATM.Accounts;
 
 /** A withdrawable account can be withdraw amount of money*/
 public interface Withdrawable {
-      void withdraw(double amount);
+      void withdraw(int amount);
 }
